airclonebackend/pyproject.toml
[tool.poetry]
name = "airbnb-clone"
version = "0.1.0"
description = ""
authors = ["rornfldla <121216565+rornfldla@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
beautifulsoup4 = "^4.12.3"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
